TCR24CS056 - Rohan Shyam/server/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
numpy==2.4.2
packaging==26.0
sqlparse==0.5.5
state-transition-diagrams==1.0.1
